--- dialog/en-us/alarm.ampm.dialog
@@ -1,3 +1,0 @@
-No problem. Shall I set it for A.M. or P.M.?
-Should I set this for A.M. or P.M.?
-Of Course. Do you want to set this for A.M. or P.M.?
