+++ dialog/en-us/am.list
@@ -1,0 +1,4 @@
+a.m.
+am
+a.m
+morning
