+++ dialog/en-us/days.list
@@ -1,0 +1,7 @@
+monday
+tuesday
+wednesday
+thursday
+friday
+saturday
+sunday
