+++ dialog/en-us/everyday.list
@@ -1,0 +1,3 @@
+day
+everyday
+daily
