--- vocab/en-us/set.morning.intent
@@ -1,3 +1,0 @@
-wake me up tomorrow morning.
-wake me up tomorrow.
-wake me up at {time} tomorrow morning.
