--- vocab/en-us/set.sunrise.intent
@@ -1,4 +1,0 @@
-Wake me up at sunrise.
-Wake me up at sunrise tomorrow.
-Can you wake me up at sunrise?
-Can you wake me up at sunrise tomorrow?
