+++ vocab/en-us/time.entity
@@ -1,3 +1,5 @@
-(#|##)(:##|)
+#
+##
+##:##
 noon
 midnight
